// packages/bot_sdk_ts/models/LeaderboardChannel.ts
/**
 * Yotarou API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class LeaderboardChannel {
    'serverId': string;
    'gameType': string;
    'channelId': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "serverId",
            "baseName": "server_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "gameType",
            "baseName": "game_type",
            "type": "string",
            "format": ""
        },
        {
            "name": "channelId",
            "baseName": "channel_id",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return LeaderboardChannel.attributeTypeMap;
    }

    public constructor() {
    }
}